Release 1.29.0

See merge request baserow/baserow!2862

+++ backend/src/baserow/contrib/builder/pages/handler.py
@@ -25,6 +25,7 @@ from baserow.contrib.builder.pages.exceptions import (
     PagePathNotUnique,
     PathParamNotDefined,
     PathParamNotInPath,
+    SharedPageIsReadOnly,
 )
 from baserow.contrib.builder.pages.models import Page
 from baserow.contrib.builder.pages.types import PagePathParams
@@ -37,7 +38,7 @@ from baserow.core.utils import ChildProgressBuilder, MirrorDict, find_unused_nam
 
 
 class PageHandler:
-    def get_page(self, page_id: int, base_queryset: QuerySet = None) -> Page:
+    def get_page(self, page_id: int, base_queryset: Optional[QuerySet] = None) -> Page:
         """
         Gets a page by ID
 
@@ -58,12 +59,30 @@ class PageHandler:
         except Page.DoesNotExist:
             raise PageDoesNotExist()
 
+    def get_shared_page(self, builder: Builder) -> Page:
+        return Page.objects.select_related("builder", "builder__workspace").get(
+            builder=builder, shared=True
+        )
+
+    def create_shared_page(self, builder: Builder) -> Page:
+        """
+        Creates the shared page of the given builder.
+
+        :param builder: The ID of the builder we want to create the shared page.
+        :return: The model instance of the shared Page
+        """
+
+        return self.create_page(
+            builder, name="__shared__", path="__shared__", shared=True
+        )
+
     def create_page(
         self,
         builder: Builder,
         name: str,
         path: str,
         path_params: PagePathParams = None,
+        shared: bool = False,
     ) -> Page:
         """
         Creates a new page
@@ -72,6 +91,8 @@ class PageHandler:
         :param name: The name of the page
         :param path: The path of the page
         :param path_params: The params of the path provided
+        :param shared: If this is the shared page. They should be only one shared page
+          per builder application.
         :return: The newly created page instance
         """
 
@@ -88,6 +109,7 @@ class PageHandler:
                 order=last_order,
                 path=path,
                 path_params=path_params,
+                shared=shared,
             )
         except IntegrityError as e:
             if "unique constraint" in e.args[0] and "name" in e.args[0]:
@@ -105,6 +127,9 @@ class PageHandler:
         :param page: The page that must be deleted
         """
 
+        if page.shared:
+            raise SharedPageIsReadOnly()
+
         page.delete()
 
     def update_page(self, page: Page, **kwargs) -> Page:
@@ -115,6 +140,9 @@ class PageHandler:
         :param kwargs: The fields that should be updated with their corresponding value
         :return: The updated page
         """
+
+        if page.shared:
+            raise SharedPageIsReadOnly()
 
         if "path" in kwargs or "path_params" in kwargs:
             path = kwargs.get("path", page.path)
@@ -159,7 +187,7 @@ class PageHandler:
         """
 
         if base_qs is None:
-            base_qs = Page.objects.filter(builder=builder)
+            base_qs = Page.objects.filter(builder=builder, shared=False)
 
         try:
             full_order = Page.order_objects(base_qs, order)
@@ -180,6 +208,9 @@ class PageHandler:
         :return: The duplicated page
         """
 
+        if page.shared:
+            raise SharedPageIsReadOnly()
+
         start_progress, export_progress, import_progress = 10, 30, 60
         progress = ChildProgressBuilder.build(progress_builder, child_total=100)
         progress.increment(by=start_progress)
@@ -197,6 +228,17 @@ class PageHandler:
 
         id_mapping = defaultdict(lambda: MirrorDict())
         id_mapping["builder_pages"] = MirrorDict()
+        id_mapping["builder_elements"] = {}
+        id_mapping["builder_workflow_actions"] = {}
+
+        shared_data_sources = DataSourceHandler().get_data_sources(
+            page=page.builder.shared_page
+        )
+        # Populate data_sources id_mapping with existing data sources as we want
+        # to keep the same Id for these.
+        id_mapping["builder_data_sources"] = {
+            ds.id: ds.id for ds in shared_data_sources
+        }
 
         new_page_clone = self.import_page(
             builder,
@@ -388,6 +430,7 @@ class PageHandler:
             order=page.order,
             path=page.path,
             path_params=page.path_params,
+            shared=page.shared,
             elements=serialized_elements,
             data_sources=serialized_data_sources,
             workflow_actions=serialized_workflow_actions,
@@ -532,13 +575,25 @@ class PageHandler:
         if "builder_pages" not in id_mapping:
             id_mapping["builder_pages"] = {}
 
-        page_instance = Page.objects.create(
-            builder=builder,
-            name=serialized_page["name"],
-            order=serialized_page["order"],
-            path=serialized_page["path"],
-            path_params=serialized_page["path_params"],
-        )
+        shared = serialized_page.get("shared", False)
+
+        if shared:
+            # The shared page has already been created at builder creation. Let's
+            # reuse that one.
+            page_instance = builder.shared_page
+            page_instance.name = serialized_page["name"]
+            page_instance.order = serialized_page["order"]
+            page_instance.path = serialized_page["path"]
+            page_instance.path_params = serialized_page["path_params"]
+        else:
+            page_instance = Page.objects.create(
+                builder=builder,
+                name=serialized_page["name"],
+                order=serialized_page["order"],
+                path=serialized_page["path"],
+                path_params=serialized_page["path_params"],
+                shared=False,
+            )
 
         id_mapping["builder_pages"][serialized_page["id"]] = page_instance.id
